Quote-Generator
The Quote Generator App is a simple web application that fetches random quotes from an API and displays them on the screen. This project aims to provide users with inspirational and thought-provoking quotes at the click of a button. The app is built using HTML, CSS, and JavaScript, making it lightweight and easy to deploy.